_raw/python/onnx_cheat/requirements-python-3.12.txt
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.0
packaging==23.1
Pillow==6.2.0
